# swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SOULMBTI API Document",
    "description": "SOULMBTI의 Swagger.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://MY_IP:8000"
    }
  ],
  "paths": {
...
    "/requests/send": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshtoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "300": {
            "description": "Multiple Choices"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "404": {
            "description": "Not Found"
          },
          "411": {
            "description": "Length Required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oppEmail": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/requests/get-received": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshtoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "300": {
            "description": "Multiple Choices"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "404": {
            "description": "Not Found"
          },
          "411": {
            "description": "Length Required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/requests/get-sended": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshtoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "300": {
            "description": "Multiple Choices"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "404": {
            "description": "Not Found"
          },
          "411": {
            "description": "Length Required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/requests/accept": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshtoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "300": {
            "description": "Multiple Choices"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "404": {
            "description": "Not Found"
          },
          "411": {
            "description": "Length Required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/requests/reject": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "accesstoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshtoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "300": {
            "description": "Multiple Choices"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "404": {
            "description": "Not Found"
          },
          "411": {
            "description": "Length Required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/requests/delete/{requestId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accesstoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshtoken",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "300": {
            "description": "Multiple Choices"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Payment Required"
          },
          "404": {
            "description": "Not Found"
          },
          "411": {
            "description": "Length Required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oppEmail": {
                    "example": "any"
                  },
                  "content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}